c.rst: Initial push of C coding standards
This is the intial configuration per Eric S, Eric M, and myself(Spenser R), that Nagios C projects should adhere to going forward. Changes are very likely as we move forward with this on projects. While we should attempt to use this as much as possible going forward, it should be understood that it may not be entirely possible at this time.